# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Lost&Found Vision API
  version: 1.0.0
servers:
  - url: https://api.example.com/v1
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Location:
      type: object
      properties:
        lat: { type: number, format: float }
        lon: { type: number, format: float }
        zone: { type: string, description: "Campus area or building" }
    Media:
      type: object
      required: [url]
      properties:
        id: { type: string }
        url: { type: string, format: uri }
        thumbnails:
          type: array
          items: { type: string, format: uri }
        safety:
          type: object
          properties:
            facesBlurred: { type: boolean, default: true }
            piiRedacted: { type: boolean, default: true }
    ItemBase:
      type: object
      required: [title, description, location]
      properties:
        title: { type: string }
        description: { type: string }
        tags:
          type: array
          items: { type: string }
        location: { $ref: '#/components/schemas/Location' }
        when:
          type: string
          format: date-time
          description: "Time last seen (lost) or time found (found)"
        photos:
          type: array
          items: { $ref: '#/components/schemas/Media' }
    LostItemCreate:
      allOf:
        - $ref: '#/components/schemas/ItemBase'
        - type: object
          properties:
            type:
              type: string
              enum: [lost]
    FoundItemCreate:
      allOf:
        - $ref: '#/components/schemas/ItemBase'
        - type: object
          properties:
            type:
              type: string
              enum: [found]
    Item:
      allOf:
        - $ref: '#/components/schemas/ItemBase'
        - type: object
          properties:
            id: { type: string }
            type: { type: string, enum: [lost, found] }
            ownerId: { type: string, nullable: true }
            finderId: { type: string, nullable: true }
            status: { type: string, enum: [active, claimed, archived], default: active }
    MatchCandidate:
      type: object
      properties:
        itemId: { type: string }
        candidateId: { type: string }
        score: { type: number, format: float }
        explanation:
          type: object
          properties:
            tags:
              type: array
              items: { type: string }
            aspects:
              type: array
              items: { type: string }
    ClaimCreate:
      type: object
      required: [itemId, candidateId, answers]
      properties:
        itemId: { type: string, description: "Lost item id (the claimant's)" }
        candidateId: { type: string, description: "Found item id being claimed" }
        answers:
          type: array
          items:
            type: object
            properties:
              question: { type: string }
              answer: { type: string }
    Claim:
      type: object
      properties:
        id: { type: string }
        itemId: { type: string }
        candidateId: { type: string }
        claimantId: { type: string }
        finderId: { type: string }
        status: { type: string, enum: [pending, verified, rejected, closed], default: pending }
        createdAt: { type: string, format: date-time }
    ClaimUpdate:
      type: object
      properties:
        action:
          type: string
          enum: [verify, reject, close]
        reason: { type: string }
    Message:
      type: object
      properties:
        id: { type: string }
        threadId: { type: string }
        senderId: { type: string }
        body: { type: string }
        createdAt: { type: string, format: date-time }
    Thread:
      type: object
      properties:
        id: { type: string }
        claimId: { type: string }
        participants:
          type: array
          items: { type: string }
        lastMessageAt: { type: string, format: date-time }
    Error:
      type: object
      properties:
        code: { type: string }
        message: { type: string }

paths:
  /items/lost:
    post:
      summary: Create a lost item report
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LostItemCreate' }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Item' }
        "400": { description: Bad Request, content: { application/json: { schema: { $ref: '#/components/schemas/Error' }}}}

  /items/found:
    post:
      summary: Create a found item post
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FoundItemCreate' }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Item' }

  /items/{id}:
    get:
      summary: Get an item by id
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Item
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Item' }
        "404": { description: Not Found }

  /items:
    get:
      summary: Search items
      parameters:
        - name: type
          in: query
          schema: { type: string, enum: [lost, found] }
        - name: zone
          in: query
          schema: { type: string }
        - name: tag
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { type: string, enum: [active, claimed, archived] }
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Item' }

  /matches:
    get:
      summary: Get match candidates for an item
      parameters:
        - name: itemId
          in: query
          required: true
          schema: { type: string }
        - name: k
          in: query
          schema: { type: integer, default: 5, minimum: 1, maximum: 20 }
      responses:
        "200":
          description: Ranked candidates
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/MatchCandidate' }

  /claims:
    post:
      summary: Open a claim on a candidate match
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ClaimCreate' }
      responses:
        "201":
          description: Claim created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Claim' }

  /claims/{id}:
    get:
      summary: Get claim
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Claim
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Claim' }
        "404": { description: Not Found }
    patch:
      summary: Update claim status (finder/moderator)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ClaimUpdate' }
      responses:
        "200":
          description: Updated claim
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Claim' }
        "403": { description: Forbidden }

  /threads/{id}/messages:
    get:
      summary: List messages in a claim thread
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Message' }
    post:
      summary: Post a message to a claim thread
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [body]
              properties:
                body: { type: string }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Message' }

  /reports:
    post:
      summary: Report content (abuse/spam/doxxing)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [targetId, reason]
              properties:
                targetId: { type: string }
                reason: { type: string }
      responses:
        "202": { description: Accepted }

  /subscriptions:
    post:
      summary: Subscribe to alerts (e.g., tag='AirPods Pro', zone='Library')
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                tag: { type: string }
                zone: { type: string }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: string }
    delete:
      summary: Unsubscribe (by id)
      parameters:
        - name: id
          in: query
          required: true
          schema: { type: string }
      responses:
        "204": { description: No Content }